health: #184 moved blueprint decorator & fns into blueprint.ts
