+++ packages/core/src/index.ts
@@ -1,4 +1,3 @@
-export * from "./lib/schema/blueprint-instance";
 export * from "./lib/entity/data-structures/complex-key-map";
 export * from "./lib/entity/data-structures/entity-set";
 export * from "./lib/entity/data-structures/normalized-entities";
@@ -33,16 +32,18 @@ export * from "./lib/query/merge-query.fn";
 export * from "./lib/query/query";
 export * from "./lib/query/reduce-queries.fn";
 export * from "./lib/query/reduce-query.fn";
+export * from "./lib/schema/blueprint";
+export * from "./lib/schema/blueprint-instance";
 export * from "./lib/schema/blueprint-property";
-export * from "./lib/schema/functions/build-default-index-name.fn";
 export * from "./lib/schema/entity-schema";
+export * from "./lib/schema/entity-schema-catalog";
 export * from "./lib/schema/entity-schema-index";
 export * from "./lib/schema/entity-schema-key";
 export * from "./lib/schema/entity-schema-property";
 export * from "./lib/schema/entity-schema-relation";
+export * from "./lib/schema/functions/build-default-index-name.fn";
 export * from "./lib/schema/property-value/array-schema";
 export * from "./lib/schema/property-value/primitive-schema";
-export * from "./lib/schema/entity-schema-catalog";
 export * from "./lib/schema/schema.interface";
 export * from "./lib/schema/unused/entity-space-schema";
 export * from "./lib/schema/unused/open-api-schema";
